refactor: change folder name to pageSimpleActions
