+++ src/components/app/pageSimpleActions/ctas/us/dayOfAction.tsx
@@ -1,7 +1,7 @@
 import { UserActionType } from '@prisma/client'
 
-import { getDefaultReferCTA } from '@/components/app/pageSimpleCampaign/ctas/commons/refer'
-import { getDefaultTweetCTA } from '@/components/app/pageSimpleCampaign/ctas/commons/tweet'
+import { getDefaultReferCTA } from '@/components/app/pageSimpleActions/ctas/commons/refer'
+import { getDefaultTweetCTA } from '@/components/app/pageSimpleActions/ctas/commons/tweet'
 import { getEmailActionWrapperComponentByCampaignName } from '@/components/app/userActionFormEmailCongressperson/getWrapperComponentByCampaignName'
 import { UserActionGridCTA } from '@/components/app/userActionGridCTAs/types'
 import { SupportedCountryCodes } from '@/utils/shared/supportedCountries'
