+++ src/components/app/pageSimpleActions/ctas/us/index.tsx
@@ -1,5 +1,5 @@
-import { US_DAY_OF_ACTION_CTAS } from '@/components/app/pageSimpleCampaign/ctas/us/dayOfAction'
-import { SimpleActionsCampaignName } from '@/components/app/pageSimpleCampaign/types'
+import { US_DAY_OF_ACTION_CTAS } from '@/components/app/pageSimpleActions/ctas/us/dayOfAction'
+import { SimpleActionsCampaignName } from '@/components/app/pageSimpleActions/types'
 import { UserActionGridCTA } from '@/components/app/userActionGridCTAs/types'
 
 export const US_SIMPLE_PAGE_CAMPAIGN_CTAS: Record<SimpleActionsCampaignName, UserActionGridCTA> = {
