+++ src/components/app/pageSimpleActions/grid.tsx
@@ -2,9 +2,9 @@
 
 import { notFound } from 'next/navigation'
 
-import { getCampaignUserActionCTAs } from '@/components/app/pageSimpleCampaign/ctas'
-import { useActions } from '@/components/app/pageSimpleCampaign/hooks/useActions'
-import { SimpleActionsCampaignName } from '@/components/app/pageSimpleCampaign/types'
+import { getCampaignUserActionCTAs } from '@/components/app/pageSimpleActions/ctas'
+import { useActions } from '@/components/app/pageSimpleActions/hooks/useActions'
+import { SimpleActionsCampaignName } from '@/components/app/pageSimpleActions/types'
 import { UserActionGridCTA } from '@/components/app/userActionGridCTAs/components/userActionGridCTA'
 import { SupportedCountryCodes } from '@/utils/shared/supportedCountries'
